Test known for memory leaks
When running read_bounds or read_dataframe each GDAL feature is not
freed and leaks memory. As this leak is outside Python objects it is not
represented/measurable via the heap and causes major issues with Dask.

This introduces memray to allow testing of memory usage (Python and/or
native).